+++ ci/macos-install-build-dependencies.sh
@@ -20,6 +20,7 @@ brew install automake || brew upgrade automake
 brew install cmake || brew upgrade cmake
 brew install bison || brew upgrade bison
 brew install devil || brew upgrade devil
+brew install gd || brew upgrade gd
 brew install gtk+ || brew upgrade gtk+
 brew install gts || brew upgrade gts
 
@@ -63,8 +64,8 @@ brew install libxaw || brew upgrade libxaw
 
 # install MacPorts for libANN
 curl --retry 3 --location --no-progress-meter -O \
-  https://github.com/macports/macports-base/releases/download/v2.7.2/MacPorts-2.7.2-12-Monterey.pkg
-sudo installer -package MacPorts-2.7.2-12-Monterey.pkg -target /
+  https://github.com/macports/macports-base/releases/download/v2.9.3/MacPorts-2.9.3-13-Ventura.pkg
+sudo installer -package MacPorts-2.9.3-13-Ventura.pkg -target /
 export PATH=/opt/local/bin:${PATH}
 
 # lib/mingle dependency