+++ cmd/tools/gv2gxl.c
@@ -486,7 +486,7 @@ static void writeSubgs(gxlstate_t * stp, Agraph_t * g, FILE * gxlFile)
     }
 }
 
-static bool writeEdgeName(Agedge_t *e, FILE *gxlFile) {
+static void writeEdgeName(Agedge_t *e, FILE *gxlFile) {
     char *p = agnameof(e);
     if (!(EMPTY(p))) {
 	tabover(gxlFile);
@@ -497,20 +497,15 @@ static bool writeEdgeName(Agedge_t *e, FILE *gxlFile) {
 	fprintf(gxlFile, "</string>\n");
 	tabover(gxlFile);
 	fprintf(gxlFile, "\t</attr>\n");
-	return true;
     }
-    return false;
 }
 
 
 static void
 writeNondefaultAttr(void *obj, FILE * gxlFile, Dict_t * defdict)
 {
-    int cnt = 0;
-
     if (AGTYPE(obj) == AGINEDGE || AGTYPE(obj) == AGOUTEDGE) {
-	if (writeEdgeName(obj, gxlFile))
-	    cnt++;
+	writeEdgeName(obj, gxlFile);
     }
     Agattr_t *data = agattrrec(obj);
     if (data) {