+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Catch2 v2 to Catch2 v3. Support for Catch2 v2 has been removed. #2465
 - The Windows release artifacts are now built on Windows 2022 and Visual Studio
   2022 17 and toolset v143.
+- The macOS release artifacts are now built on macOS 13 (Ventura).
 
 ### Fixed